+++ app/controllers/admin/danger_zone_controller.rb
@@ -2,13 +2,11 @@ class Admin::DangerZoneController < AdminController
   before_action :set_owner
 
   def drop_contacts
-    @owner.contacts.destroy_all
-    redirect_to admin_owner_path(@owner.mkey), alert: 'All contacts was destroyed'
+    handle_admin_action(Admin::Owners::DropContacts.new(@owner), admin_owner_path(@owner.mkey))
   end
 
   def drop_notifications
-    @owner.notifications.destroy_all
-    redirect_to admin_owner_path(@owner.mkey), alert: 'All notifications was destroyed'
+    handle_admin_action(Admin::Owners::DropNotifications.new(@owner), admin_owner_path(@owner.mkey))
   end
 
   def clear_statuses