+++ app/services/admin/owners/drop_notifications.rb
@@ -1,0 +1,6 @@
+class Admin::Owners::DropNotifications < Admin::Owners
+  def do
+    @owner.notifications.destroy_all
+    [true, 'All notifications was destroyed']
+  end
+end